# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API to fetch hotel reviews and ratings.",
        "title": "Hotel Review API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/hotels/{hotel_id}/reviews": {
            "get": {
                "description": "Returns average rating and paginated reviews for a hotel",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Get hotel reviews and overall rating",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hotel ID",
                        "name": "hotel_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Reviews per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ReviewResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AggregatedHotelReview": {
            "type": "object",
            "properties": {
                "averageRating": {
                    "type": "number"
                },
                "hotelID": {
                    "type": "integer"
                },
                "hotelName": {
                    "type": "string"
                },
                "reviewCount": {
                    "type": "integer"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.ReviewDetail": {
            "type": "object",
            "properties": {
                "country_name": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "review_date": {
                    "type": "string"
                },
                "review_group_name": {
                    "type": "string"
                },
                "review_text": {
                    "type": "string"
                },
                "review_title": {
                    "type": "string"
                },
                "room_type_name": {
                    "type": "string"
                }
            }
        },
        "models.ReviewResponse": {
            "type": "object",
            "properties": {
                "hotel": {
                    "$ref": "#/definitions/models.AggregatedHotelReview"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ReviewDetail"
                    }
                }
            }
        }
    }
}